+++ ember_ml/wave/memory/sphere_overlap.py
@@ -53,10 +53,10 @@ class SphereState:
     def __post_init__(self):
         """Convert states to numpy arrays and validate."""
         if not isinstance(self.fast_state, np.ndarray):
-            self.fast_state = np.array(self.fast_state, dtype=np.float64)
+            self.fast_state = np.array(self.fast_state, dtype=np.float32)
             
         if not isinstance(self.slow_state, np.ndarray):
-            self.slow_state = np.array(self.slow_state, dtype=np.float64)
+            self.slow_state = np.array(self.slow_state, dtype=np.float32)
             
         if self.fast_state.shape != (4,):
             raise ValueError("Fast state must be a 4D vector")
